### 2016-06-15 #28 @whedon
Submission: Phylogemetric: A Python library for calculating phylogenetic network metrics
**Submitting author:** @SimonGreenhill (<a href="http://orcid.org/0000-0001-7832-6156">Simon Greenhill</a>)
**Repository:** <a href="https://github.com/SimonGreenhill/phylogemetric" target ="_blank">https://github.com/SimonGreenhill/phylogemetric</a>
**Version:** v1.0.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.55663" target ="_blank">http://dx.doi.org/10.5281/zenodo.55663</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.55663">software archive</a> resolve?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
