### 2016-06-16 #28 @krother
Checked both doi-less references. They in fact have no doi.
