### 2016-06-16 #28 @krother
This is a small but very neat piece of software. It works smoothly and I understood everything (the Eigen paper on quartet analysis/statistical geometry was one of my favourites as a student). I have a couple of suggestions:

- One line on how to install the program (python setup.py install works for me) would be fine.
- One example (plus adding a small Nexus file to the repository) would make it a lot easier to try the program.
- Please use a standard license. There are good reasons not to write your own legal text (https://www.youtube.com/watch?v=H_X9TK_VJak)
- There is no community guideline. In my opinion this is not crucial because the project is small and has an awesome test suite (running tests could be mentioned).

From an engineering point of view, I find it very convincing. Good job!
