### 2016-06-16 1c045a0
Create create_random_nexus.py
Created random nexus files to test the program on a larger dataset.
