### 2016-06-16 #28 @krother
Another idea: since the Holland paper is available in full-text it could be hyperlinked from the README.
