### 2016-06-16 #28 @danielskatz
@arfon, it's close, but not quite standard.
