### 2016-06-16 #28 @arfon
> @arfon, it's close, but not quite standard.

@danielskatz - the only difference I see here is that they've substituted 'copyright holder' for 'Treemaker' on this line:

3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.

Am I missing something?
