### 2016-06-16 #28 @danielskatz
No, but that seems to me to be a potentially important difference, and one that doesn't have any benefit.
