### 2016-06-16 #28 @arfon
> No, but that seems to me to be a potentially important difference, and one that doesn't have any benefit.

👍 ok. I agree that for clarity this should be switched back to the standard language. @SimonGreenhill - would you mind switching the language back to the license text here: https://opensource.org/licenses/BSD-3-Clause
