### 2016-06-17 #28 @SimonGreenhill
Hi all, thanks for the comments/suggestions, I'll make them ASAP.

I had intended a generic BSD-3 clause, so not sure what happened (perhaps I did a dumb search and replace that messed something up -- TreeMaker  is another package I wrote). Will revert to BSD-3
