### 2016-06-17 b631403
adding install info
