+++ README.md
@@ -6,6 +6,13 @@ A python library for calculating delta score (Holland et al. 2002) and Q-Residua
 [![Coverage Status](https://coveralls.io/repos/SimonGreenhill/phylogemetric/badge.svg?branch=master&service=github)](https://coveralls.io/github/SimonGreenhill/phylogemetric?branch=master)
 [![DOI](https://zenodo.org/badge/22704/SimonGreenhill/phylogemetric.svg)](https://zenodo.org/badge/latestdoi/22704/SimonGreenhill/phylogemetric)
 
+## Installation:
+
+Installation is only a pip install away:
+
+```shell
+pip install phylogemetric
+```
 
 ## Usage: Command line
 
@@ -106,7 +113,6 @@ m.score()
 m.pprint()
 
 ```
-
 
 ## Requirements:
 
