### 2016-06-17 f7843be
adding CONTRIBUTING.md
