+++ CONTRIBUTING.md
@@ -1,0 +1,32 @@
+# Contributing
+
+We'd love to extend phylogemetric to enable new methods and analyses. Please add issues and 
+suggestions to the GitHub Issues page. 
+
+If you want to get your hands dirty and change the code, then:
+
+Fork the repo:
+
+    git clone git@github.com:SimonGreenhill/phylogemetric.git
+
+Install the test framework py.test
+
+    pip install pytest pytest-cov
+
+Make sure tests pass:
+
+    make test
+
+Make your change.
+
+Write tests.
+
+Follow the PEP8 [https://www.python.org/dev/peps/pep-0008/][style guide].
+
+Make the tests pass:
+
+    make test
+
+Submit a pull request.
+
+
