### 2016-06-17 9805139
rename rst file to md for consistency
