+++ DESCRIPTION.md
@@ -1,0 +1,6 @@
+Phylogemetric
+-------------
+
+A Python library for calculating the delta score (Holland et al. 2002) and the Q-Residual (Gray et al. 2010) from
+phylogenetic nexus files.
+
