### 2016-06-17 e7c7028
adding links to README
