+++ README.md
@@ -1,6 +1,6 @@
 # phylogemetric
 
-A python library for calculating delta score (Holland et al. 2002) and Q-Residual (Gray et al. 2010) for phylogenetic data.
+A python library for calculating delta score ([Holland et al. 2002](http://mbe.oxfordjournals.org/content/19/12/2051.full)) and Q-Residual ([Gray et al. 2010](http://dx.doi.org/10.1098/rstb.2010.0162)) for phylogenetic data.
 
 [![Build Status](https://travis-ci.org/SimonGreenhill/phylogemetric.svg?branch=master)](https://travis-ci.org/SimonGreenhill/phylogemetric)
 [![Coverage Status](https://coveralls.io/repos/SimonGreenhill/phylogemetric/badge.svg?branch=master&service=github)](https://coveralls.io/github/SimonGreenhill/phylogemetric?branch=master)
