+++ paper/paper.bib
@@ -7,7 +7,9 @@
     pmid = {21041216},
     title = {{On the shape and fabric of human history.}},
     volume = {365},
-    year = {2010}
+    year = {2010},
+    url = {http://dx.doi.org/10.1098/rstb.2010.0162},
+    doi = {10.1098/rstb.2010.0162}
 }
 
 @article{Holland2002,
@@ -17,7 +19,8 @@
     pages = {2051--2059},
     title = {{[delta] Plots: A tool for analyzing phylogenetic distance data}},
     volume = {19},
-    year = {2002}
+    year = {2002},
+    url = {http://mbe.oxfordjournals.org/content/19/12/2051.full}
 }
 
 @article{Nexus,
