### 2016-06-18 #28 @arfon
> - Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.

@danielskatz I've been investigating how GitHub handles the templating of the BSD 3-clause license. It looks like we swap out `copyright holder` for the name of the repository: see [here](https://github.com/arfon/test-repository/blob/master/LICENSE#L14) for example.

As @SimonGreenhill entinoed, in this case it looks like `TreeMaker` is the name of another package but it seems like we should be tolerant/allow people to replace `copyright holder` with an actual name here.

/ cc @mlinksva incase he has a chance to give feedback here :grin:
