### 2016-06-18 #28 @danielskatz
@arfon I disagree with you about the templating.  In clause 3, I'm fairly sure "copyright holder" is part of the license text, not a template variable to be filled in.  See https://opensource.org/licenses/BSD-3-Clause   This makes it clear (to me, at least) that the only variation in the license should be in the first line.
