### 2016-06-18 #28 @arfon
@danielskatz - sorry, should have clarified: I wasn't trying to say this was the definitely the 'correct' behaviour, rather just try and understand why/how 'copyright holder' has been swapped out for something else. It seems like this is probably GitHub's fault.

It does seems like this treating this as a template variable might be an error on GitHub's part here hence my ping of @mlinksva
