### 2016-06-18 #28 @mlinksva
There's not a single right answer, but I agree less templated fields is good. This has been recently fixed in https://github.com/github/choosealicense.com/pull/415 (and see discussion there for some of the history of the license) and has been vendored into https://github.com/benbalter/licensee/releases/tag/v8.2.0 which will make its way into the github.com license chooser when it gets that or a later version of licensee. This will result in fewer projects with license files with clause 3 templated, but there's nothing wrong with license files that do have that clause templated.
