### 2016-06-18 #28 @danielskatz
For JOSS, If there's a standard, and we get submissions that don't use the standard, we shouldn't accept them.

The license is still valid, and if that's what the creators want to use, that's fine, but in the context of JOSS, I don't think this is what we want.
