### 2016-06-18 #28 @arfon
@danielskatz - agreed.

@SimonGreenhill - I think you offered already but just to confirm - please switch this out to the standard license text :smile_cat:
