### 2016-06-20 #28 @SimonGreenhill
@arfon - done. Sorry for the confusion! I'll push an update soon that fixes the license issue, and the review comments by @krother
