### 2016-06-20 cd7cc09
refactoring with flake8
