+++ phylogemetric/tests/test_delta.py
@@ -52,4 +52,3 @@ class Test_DeltaScoreMetric_Complex(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-    
