+++ phylogemetric/tests/test_metric.py
@@ -1,4 +1,7 @@
+import re
+import sys
 import unittest
+from io import StringIO
 from phylogemetric.metric import Metric
 from phylogemetric.tests.data_simple import MATRIX
 
@@ -37,6 +40,7 @@ class TestMetric(unittest.TestCase):
     # tests that require data
     def setUp(self):
         self.metric = Metric(MATRIX)
+        self.metric.score()
     
     def test_get_dist_same_taxon(self):
         assert self.metric.get_dist('A', 'A', [0, 1], [0, 1]) == 0.0
@@ -46,4 +50,19 @@ class TestMetric(unittest.TestCase):
         assert m.qscores is not None
         m.qscores = None
         m.score()
-        
+    
+    def test_pprint(self):
+        stdout = sys.stdout
+        sys.stdout = StringIO()
+        self.metric.pprint()
+        out = sys.stdout.getvalue()
+        sys.stdout.close()
+        sys.stdout = stdout
+        # all scores will be 0 as `Metric` doesn't implement a distance
+        expected = sorted(MATRIX.keys())
+        obtained = sorted(re.findall(r"""([A-E]{1})\s+""", out))
+        self.assertEqual(expected, obtained)
+
+
+if __name__ == '__main__':
+    unittest.main()
