### 2016-06-20 #28 @SimonGreenhill
Ok, all changes suggested by @krother have been made, and the LICENSE file has been updated. Many thanks for all your suggestions.
