+++ README.md
@@ -5,6 +5,8 @@ A python library for calculating delta score ([Holland et al. 2002](http://mbe.o
 [![Build Status](https://travis-ci.org/SimonGreenhill/phylogemetric.svg?branch=master)](https://travis-ci.org/SimonGreenhill/phylogemetric)
 [![Coverage Status](https://coveralls.io/repos/SimonGreenhill/phylogemetric/badge.svg?branch=master&service=github)](https://coveralls.io/github/SimonGreenhill/phylogemetric?branch=master)
 [![DOI](https://zenodo.org/badge/22704/SimonGreenhill/phylogemetric.svg)](https://zenodo.org/badge/latestdoi/22704/SimonGreenhill/phylogemetric)
+[![License](https://img.shields.io/pypi/l/phylogemetric.svg)](https://github.com/SimonGreenhill/phylogemetric/blob/master/LICENSE)
+
 
 ## Installation:
 
