### 2016-06-20 3abf9e1
removing unneeded comment
