### 2016-06-20 a6236fa
adding test for create_nex
