### 2016-06-20 32eb65a
no cover a defensive import
