+++ phylogemetric/metric.py
@@ -93,5 +93,5 @@ class Metric(object):
     def pprint(self):
         max_len = max([len(_) for _ in self.matrix])
         for taxon in sorted(self.scores):
-            print("%s\t%0.4f" % (taxon.ljust(max_len + 1), self.scores[taxon]))
+            print("%s\t%f" % (taxon.ljust(max_len + 1), self.scores[taxon]))
 
