### 2016-06-20 #28 @SimonGreenhill
Sure, how's [this](https://github.com/SimonGreenhill/phylogemetric/commit/a030bd057a40b79ce1a3d10cef52168f7f5ec5fe)?
