### 2016-06-20 #28 @arfon
> Sure, how's this?

Looks good to me @SimonGreenhill.

Thanks for the rapid review here @krother. I've updated the checklist to make it clear that the license is now updated.

@SimonGreenhill - your DOI is http://dx.doi.org/10.21105/joss.00028 🎉 🚀 :boom:
